lt-qcom-debian-images-arm64: fix cleanup scripts
In the previous rework of builders.sh , it was split into multiple subscripts,
so that we could reuse them. However the way the cleanup traps were done at the
end of the first subscripts, the entire workspace was wiped out, so the next
subscript failed. So let's defer the workspace cleanup until the last piece
(publishing), and only do early cleanup in case of failure.

+++ lt-qcom-debian-images-arm64/builders.sh
@@ -6,6 +6,9 @@ trap cleanup_exit INT TERM EXIT
 
 cleanup_exit()
 {
+    # cleanup here, only in case of error in this script
+    # normal cleanup deferred to later
+    [ $? = 0 ] && exit;
     cd ${WORKSPACE}
     sudo kpartx -dv out/${VENDOR}-${OS_FLAVOUR}-*.sd.img || true
     sudo git clean -fdxq
